+++ util/chplenv/chplenv.py
@@ -33,3 +33,4 @@ import chpl_python_version
 import compiler_utils
 import overrides
 import utils
+import compile_link_args_utils
